Refactor Player:UpdateCards() to accept a gearSelector parameter
This commit refactors the Player:UpdateCards() function to accept a gearSelector parameter. This allows for more flexibility in selecting which gear cards to update. The gearSelector function is used to filter the gear cards based on specific criteria. If no gearSelector is provided, the default behavior is to select gear cards with the "cursed" stat.